+++ Funcionarios.php
@@ -95,10 +95,11 @@ $result = $conexao->query($sql);
                         <th scope="col">Senha</th>
                         <th scope="col">Telefone</th>
                         <th scope="col">sexo</th>
+                        <th scope="col">Especies</th>
                         <th scope="col">Data de Nascimento</th>
                         <th scope="col">Cidade</th>
                         <th scope="col">Estado</th>
-                        <th scope="col">Endereço</th>
+                        <th scope="col">Endereço</th>  
                         <th scope="col">...</th>
                     </tr>
                 </thead>
@@ -114,10 +115,12 @@ $result = $conexao->query($sql);
                         echo "<td>" . $user_data['senha'] . "</td>";
                         echo "<td>" . $user_data['telefone'] . "</td>";
                         echo "<td>" . $user_data['sexo'] . "</td>";
+                        echo "<td>" . $user_data['especie'] . "</td>";
                         echo "<td>" . $user_data['data_nascimento'] . "</td>";
                         echo "<td>" . $user_data['cidade'] . "</td>";
                         echo "<td>" . $user_data['estado'] . "</td>";
                         echo "<td>" . $user_data['endereco'] . "</td>";
+                       
                         echo "<td>
                             <a class='btn1 btn-sm btn-primary' href='Edit/edit.php?id=$user_data[id]'>
                             <svg xmlns='http://www.w3.org/2000/svg' width='16' height='16' fill='currentColor' class='bi bi-pencil-square' viewBox='0 0 16 16'>
@@ -125,7 +128,7 @@ $result = $conexao->query($sql);
                             <path fill-rule='evenodd' d='M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5z'/>
                             </svg> 
                             </a>    
-                            <a class='btn2 btn-sm btn-danger' href='Delete/delete.php?id=$user_data[id]'>
+                            <a class='btn2 btn-sm btn-danger' href='Delete/delete.php?id=". $user_data["id"] . "' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\");'>
                             <svg xmlns='http://www.w3.org/2000/svg' width='16' height='16' fill='currentColor' class='bi bi-trash-fill' viewBox='0 0 16 16'>
                             <path d='M2.5 1a1 1 0 0 0-1 1v1a1 1 0 0 0 1 1H3v9a2 2 0 0 0 2 2h6a2 2 0 0 0 2-2V4h.5a1 1 0 0 0 1-1V2a1 1 0 0 0-1-1H10a1 1 0 0 0-1-1H7a1 1 0 0 0-1 1zm3 4a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-1 0v-7a.5.5 0 0 1 .5-.5M8 5a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-1 0v-7A.5.5 0 0 1 8 5m3 .5v7a.5.5 0 0 1-1 0v-7a.5.5 0 0 1 1 0'/>
                             </svg>
